#970 finish cleaning paths
