+++ src/hyperion/parameters/external_parameters.py
@@ -33,7 +33,7 @@ def from_json(json_params: str):
 
 
 def from_file(
-    json_filename: str = "tests/parameter_json_files/test_parameter_defaults.json",
+    json_filename: str = "tests/test_data/parameter_json_files/test_parameter_defaults.json",
 ):
     with open(json_filename) as f:
         return from_json(f.read())
