+++ tests/system_tests/external_interaction/test_write_rotation_nexus.py
@@ -19,14 +19,14 @@ from hyperion.parameters.plan_specific.rotation_scan_internal_params import (
 )
 
 TEST_EXAMPLE_NEXUS_FILE = Path("ins_8_5.nxs")
-TEST_DIRECTORY = Path("src/hyperion/external_interaction/unit_tests/test_data/")
+TEST_DIRECTORY = Path("tests/test_data/scratch")
 TEST_FILENAME = "rotation_scan_test_nexus"
 
 
 @pytest.fixture
 def test_params():
     param_dict = from_file(
-        "src/hyperion/parameters/tests/test_data/good_test_rotation_scan_parameters.json"
+        "tests/test_data/parameter_json_files/good_test_rotation_scan_parameters.json"
     )
     param_dict["hyperion_params"]["detector_params"][
         "directory"
