+++ tests/test_data/hyperion_parameters.json
@@ -15,7 +15,7 @@
         "num_images_per_trigger": 1,
         "num_triggers": 60,
         "use_roi_mode": false,
-        "det_dist_to_beam_converter_path": "tests/unit_tests/hyperion/test_lookup_table.txt",
+        "det_dist_to_beam_converter_path": "tests/test_data/test_lookup_table.txt",
         "trigger_mode": 2,
         "detector_size_constants": "EIGER2_X_16M",
         "beam_xy_converter": null,
