+++ tests/test_data/parameter_json_files/bad_test_parameters_wrong_version.json
@@ -1,18 +1,22 @@
 {
-    "params_version": "4.0.0",
+    "params_version": "0.0.4",
     "hyperion_params": {
-        "beamline": "BL03I",
-        "insertion_prefix": "SR03I",
+        "beamline": "BL03S",
+        "insertion_prefix": "SR03S",
         "detector": "EIGER2_X_16M",
         "zocalo_environment": "dev_artemis",
-        "experiment_type": "rotation_scan",
+        "experiment_type": "grid_scan",
         "detector_params": {
-            "current_energy_ev": 12700,
-            "directory": "/dls/i03/data/2023/cm33866-3/rotation_scan_test",
-            "prefix": "rotation_scan_test",
+            "current_energy_ev": 100,
+            "exposure_time": 0.1,
+            "directory": "/tmp",
+            "prefix": "file_name",
             "run_number": 0,
+            "detector_distance": 100.0,
+            "omega_start": 0.0,
+            "num_images": 50,
             "use_roi_mode": false,
-            "det_dist_to_beam_converter_path": "tests/unit_tests/hyperion/test_lookup_table.txt"
+            "det_dist_to_beam_converter_path": "tests/test_data/test_lookup_table.txt"
         },
         "ispyb_params": {
             "visit_path": "/tmp/cm31105-4/",
@@ -56,13 +60,17 @@
         }
     },
     "experiment_params": {
-        "rotation_axis": "omega",
-        "rotation_angle": 180.0,
-        "omega_start": 0.0,
-        "exposure_time": 0.01,
-        "detector_distance": 300.0,
-        "rotation_increment": 0.1,
-        "rotation_direction": "NEGATIVE",
-        "shutter_opening_time_s": 0.6
+        "x_steps": 5,
+        "y_steps": 10,
+        "z_steps": 2,
+        "x_step_size": 0.1,
+        "y_step_size": 0.1,
+        "z_step_size": 0.1,
+        "dwell_time_ms": 2,
+        "x_start": 0.0,
+        "y1_start": 0.0,
+        "y2_start": 0.0,
+        "z1_start": 0.0,
+        "z2_start": 0.0
     }
 }
