+++ tests/test_data/parameter_json_files/good_test_pin_centre_then_xray_centre_parameters.json
@@ -4,15 +4,15 @@
         "beamline": "BL03S",
         "insertion_prefix": "SR03S",
         "detector": "EIGER2_X_16M",
-        "zocalo_environment": "dev_artemis",
-        "experiment_type": "full_grid_scan",
+        "zocalo_environment": "devrmq",
+        "experiment_type": "pin_tip_centre_then_xray_centre",
         "detector_params": {
             "current_energy_ev": 100,
             "directory": "/tmp",
             "prefix": "file_name",
             "run_number": 0,
             "use_roi_mode": false,
-            "det_dist_to_beam_converter_path": "tests/unit_tests/hyperion/test_lookup_table.txt"
+            "det_dist_to_beam_converter_path": "tests/test_data/test_lookup_table.txt"
         },
         "ispyb_params": {
             "visit_path": "/tmp/cm31105-4/",
@@ -41,6 +41,8 @@
         "exposure_time": 0.1,
         "detector_distance": 100.0,
         "omega_start": 0.0,
-        "grid_width_microns": 151
+        "tip_offset_microns": 108.9,
+        "grid_width_microns": 290.6,
+        "oav_centring_file": "tests/test_data/test_OAVCentring.json"
     }
 }
