+++ tests/test_data/parameter_json_files/good_test_rotation_scan_parameters.json
@@ -12,7 +12,7 @@
             "prefix": "file_name",
             "run_number": 0,
             "use_roi_mode": false,
-            "det_dist_to_beam_converter_path": "tests/unit_tests/hyperion/test_lookup_table.txt"
+            "det_dist_to_beam_converter_path": "tests/test_data/test_lookup_table.txt"
         },
         "ispyb_params": {
             "visit_path": "/tmp/cm31105-4/",
@@ -59,6 +59,11 @@
         "rotation_axis": "omega",
         "rotation_angle": 180.0,
         "omega_start": 0.0,
+        "phi_start": 0.0,
+        "chi_start": 0,
+        "x": 1.0,
+        "y": 2.0,
+        "z": 3.0,
         "exposure_time": 0.1,
         "detector_distance": 100.0,
         "rotation_increment": 0.1,
