+++ tests/test_data/parameter_json_files/live_test_rotation_params_move_xyz.json
@@ -12,7 +12,7 @@
             "prefix": "rotation_scan_test",
             "run_number": 0,
             "use_roi_mode": false,
-            "det_dist_to_beam_converter_path": "tests/unit_tests/hyperion/test_lookup_table.txt"
+            "det_dist_to_beam_converter_path": "tests/test_data/test_lookup_table.txt"
         },
         "ispyb_params": {
             "visit_path": "/tmp/cm31105-4/",
