+++ tests/test_data/parameter_json_files/test_parameters.json
@@ -4,7 +4,7 @@
         "beamline": "BL03S",
         "insertion_prefix": "SR03S",
         "detector": "EIGER2_X_16M",
-        "zocalo_environment": "devrmq",
+        "zocalo_environment": "dev_artemis",
         "experiment_type": "flyscan_xray_centre",
         "detector_params": {
             "current_energy_ev": 100,
@@ -12,8 +12,7 @@
             "prefix": "file_name",
             "run_number": 0,
             "use_roi_mode": false,
-            "det_dist_to_beam_converter_path": "tests/unit_tests/hyperion/test_lookup_table.txt",
-            "num_triggers": 1
+            "det_dist_to_beam_converter_path": "tests/test_data/test_lookup_table.txt"
         },
         "ispyb_params": {
             "visit_path": "/tmp/cm31105-4/",
