+++ tests/test_data/scratch/README
@@ -1,0 +1,1 @@
+This directory exists for tests to temporarily write files. Anything in here except this readme can be safely deleted.
