+++ tests/test_data/test_display.configuration
