+++ tests/test_data/test_jCameraManZoomLevels.xml
