+++ tests/unit_tests/device_setup_plans/test_setup_oav.py
@@ -16,9 +16,9 @@ from hyperion.device_setup_plans.setup_oav import (
     pre_centring_setup_oav,
 )
 
-ZOOM_LEVELS_XML = "tests/test_data/jCameraManZoomLevels.xml"
-OAV_CENTRING_JSON = "tests/test_data/OAVCentring.json"
-DISPLAY_CONFIGURATION = "tests/test_data/display.configuration"
+ZOOM_LEVELS_XML = "tests/test_data/test_jCameraManZoomLevels.xml"
+OAV_CENTRING_JSON = "tests/test_data/test_OAVCentring.json"
+DISPLAY_CONFIGURATION = "tests/test_data/test_display.configuration"
 
 
 @pytest.fixture
