--- tests/unit_tests/experiment_plans/test_data/display.configuration
@@ -1,42 +1,0 @@
-zoomLevel = 1.0
-crosshairX = 477
-crosshairY = 359
-topLeftX = 383
-topLeftY = 253
-bottomRightX = 410
-bottomRightY = 278
-zoomLevel = 2.5
-crosshairX = 493
-crosshairY = 355
-topLeftX = 340
-topLeftY = 283
-bottomRightX = 388
-bottomRightY = 322
-zoomLevel = 5.0
-crosshairX = 517
-crosshairY = 350
-topLeftX = 268
-topLeftY = 326
-bottomRightX = 354
-bottomRightY = 387
-zoomLevel = 7.5
-crosshairX = 549
-crosshairY = 347
-topLeftX = 248
-topLeftY = 394
-bottomRightX = 377
-bottomRightY = 507
-zoomLevel = 10.0
-crosshairX = 613
-crosshairY = 344
-topLeftX = 2
-topLeftY = 489
-bottomRightX = 206
-bottomRightY = 630
-zoomLevel = 15.0
-crosshairX = 693
-crosshairY = 339
-topLeftX = 1
-topLeftY = 601
-bottomRightX = 65
-bottomRightY = 767
