+++ tests/unit_tests/external_interaction/callbacks/test_rotation_callbacks.py
@@ -36,7 +36,7 @@ from hyperion.parameters.plan_specific.rotation_scan_internal_params import (
 def params():
     return RotationInternalParameters(
         **from_file(
-            "src/hyperion/parameters/tests/test_data/good_test_rotation_scan_parameters.json"
+            "tests/test_data/parameter_json_files/good_test_rotation_scan_parameters.json"
         )
     )
