+++ tests/unit_tests/parameters/plan_specific/test_stepped_grid_scan_internal_parameters.py
@@ -7,7 +7,7 @@ from hyperion.parameters.plan_specific.stepped_grid_scan_internal_params import 
 
 def test_stepped_grid_scan_parameters_load_from_file():
     params = external_parameters.from_file(
-        "src/hyperion/parameters/tests/test_data/good_test_stepped_grid_scan_parameters.json"
+        "tests/test_data/parameter_json_files/good_test_stepped_grid_scan_parameters.json"
     )
     internal_parameters = SteppedGridScanInternalParameters(**params)
